+++ scripts/libcxl_examples.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+set -e
+
+. libcxl.sh
+
+cxl_info_all